Update patient deletion to delete patient from associated doctor's patients list

+++ services/patientService.js
@@ -1,21 +1,22 @@
 import mongoose from 'mongoose';
 import Patient from '../middlewares/patientMiddleware';
+// import Doctor from '../middlewares/doctorMiddleware';
 import DoctorService from './doctorService';
 
 class PatientService {
-  getPatients (page = 0, size = 10) {
+  getPatients(page = 0, size = 10) {
     return Patient.find().skip(page * size).limit(size);
   }
 
-  async getPatientById (id) {
+  async getPatientById(id) {
     return Patient.findById(id);
   }
 
-  async getPatientByEmail (email) {
+  async getPatientByEmail(email) {
     return Patient.findOne({ email });
   }
 
-  async createPatient (query) {
+  async createPatient(query) {
     if (query.doctorId) {
       query.doctor = new mongoose.Types.ObjectId(query.doctorId);
       delete query.doctorId;
@@ -36,7 +37,7 @@ class PatientService {
     return patient;
   }
 
-  async updateAPatient (patient, query) {
+  async updateAPatient(patient, query) {
     // Make sure the email is unique, by checking if the email in the query
     // We override the email in the query with the email in the database
     // If not, the unique property in Patient's schema will raise an error 'duplicate key'
@@ -51,8 +52,21 @@ class PatientService {
     return updatedUser;
   }
 
-  async deleteAPatient (id) {
-    return Patient.findByIdAndDelete(id);
+  async deleteAPatient(id, doctorId) {
+    try {
+      if (id) {
+        const doctor = await DoctorService.getDoctorById(doctorId);
+        if (doctor) {
+          const query = {};
+          query.patients = doctor.patients.filter(patientId => !patientId.equals(id));
+          await DoctorService.updateADoctor(doctor, query);
+        }
+        return Patient.findByIdAndDelete(id);
+      }
+    } catch (error) {
+      console.error('Error deleting patient:', error);
+      throw new Error('Failed to delete the patient.');
+    }
   }
 }
